move scrapper inside helper and add populator code
