+++ project/helper.py
@@ -1,4 +1,7 @@
 import uuid
+import newspaper
+import youtube_dl
+import re
 
 def generate_id():
     return str(uuid.uuid1())
@@ -58,3 +61,29 @@ def generate_video(title, youtube_id, description, duration, view_count, ofGame=
         "hasTags": hasTags if hasSubs else [],
         "hasSubs": hasSubs if hasSubs else [],
     }
+
+def scrap_article(article_link):
+    article = newspaper.Article(article_link)
+    article.download()
+    article.parse()
+
+    raw_paragraph = article.text
+    paragraph_list = [p for p in raw_paragraph.split("\n") if p]
+    return {
+        "title": article.title,
+        "paragraph_list": paragraph_list
+    }
+
+def scrap_video(video_url):
+    ydl_opts = {
+        "writeautomaticsub": True,
+        "skip_download": True,
+        "quiet": True,
+    }
+    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
+        ydl.download([video_url])
+
+def extract_autosub(subtitle_path):
+    with open(subtitle_path) as i:
+        raw_text = i.read()
+        return re.findall("(\d\d:\d\d:\d\d\.\d\d\d) --> (\d\d:\d\d:\d\d\.\d\d\d).+\n(.{2,})\n", raw_text)
