+++ project/populate_schema.py
@@ -1,38 +1,26 @@
-import newspaper
-import youtube_dl
-import re
+from project import helper
+import glob
+import os
 
-def scrap_article(article_link):
-    article = newspaper.Article(article_link)
-    article.download()
-    article.parse()
+def populate_video():
+    with open("video_links") as i:
+        for link in i:
+            helper.scrap_video(link.strip())
+            subtitle_list = glob.glob("*.vtt")
+            if len(subtitle_list):
+                subtitle_path = subtitle_list[0]
+                extracted = helper.extract_autosub(subtitle_path)
+                for e in extracted:
+                    print(e)
+                    # todo: insert into graph
+                os.remove(subtitle_path)
 
-    raw_paragraph = article.text
-    paragraph_list = [p for p in raw_paragraph.split("\n") if p]
-    return {
-        "title": article.title,
-        "paragraph_list": paragraph_list
-    }
+def populate_article():
+    with open("article_links") as i:
+        for link in i:
+            result = helper.scrap_article(link.strip())
+            print(result)
+            # todo: insert into graph
 
-def scrap_video(video_url):
-    ydl_opts = {
-        "writeautomaticsub": True,
-        "skip_download": True,
-        "quiet": True,
-    }
-    with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-        ydl.download([video_url])
-
-def extract_autosub(subtitle_path):
-    with open(subtitle_path) as i:
-        raw_text = i.read()
-        return re.findall("(\d\d:\d\d:\d\d\.\d\d\d) --> (\d\d:\d\d:\d\d\.\d\d\d).+\n(.{2,})\n", raw_text)
-
-
-# print(scrap_article(('https://www.gamesradar.com/gta-5-guide/')))
-
-# scrap_video('https://www.youtube.com/watch?v=Vncf_9LLagc') # gta
-# scrap_video('https://www.youtube.com/watch?v=BaW_jenozKc') # ytdl example video
-
-# process_autosub("autosub.vtt")
-extract_autosub("autosub.vtt")
+# populate_video()
+# populate_article()
